requirements/local.txt
bcrypt
certifi
cffi
chardet
defusedxml
Django
django-admin-honeypot
django-analytical
django-appconf
django-axes
django-debug-toolbar
django-ipware
docutils
et-xmlfile
idna
isort
jdcal
oauthlib
openpyxl
psycopg
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
Unidecode
urllib3
